requirements/mminstall.txt
mmcv-full>=1.3.0,<=1.5.0
mmcls==0.13.0